+++ src/aui-datepicker/js/aui-datepicker.js
@@ -195,6 +195,10 @@ A.mix(DatePickerBase.prototype, {
         instance.alignTo(node);
         instance.clearSelection(true);
         instance.selectDates(instance.getParsedDatesFromInputValue());
+
+        instance.fire('useInputNode', {
+            node: node
+        });
     },
 
     /**